+++ src/redux/reducers/reducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
-import savePlaceReducer from './savePlace';
+import placeReducer from './placeReducer';
 import notificationReducer from './notificationReducer';
 
-const rootReducer = combineReducers({ savePlaceReducer, notificationReducer });
+const rootReducer = combineReducers({ placeReducer, notificationReducer });
 
 export default rootReducer;
